Refactor TOS acceptance condition to optimize typical code path

+++ woocommerce-services.php
@@ -380,15 +380,17 @@ if ( ! class_exists( 'WC_Connect_Loader' ) ) {
 		public function pre_wc_init() {
 			$this->load_dependencies();
 
+			$tos_accepted = WC_Connect_Options::get_option( 'tos_accepted' );
+
 			// Prevent presenting users with TOS they've already
 			// accepted in the core WC Setup Wizard or on WP.com
-			if ( get_option( 'woocommerce_setup_jetpack_opted_in' ) || WC_Connect_Jetpack::is_atomic_site() ) {
+			if ( ! $tos_accepted &&
+				( get_option( 'woocommerce_setup_jetpack_opted_in' ) || WC_Connect_Jetpack::is_atomic_site() )
+			) {
 				WC_Connect_Options::update_option( 'tos_accepted', true );
 				delete_option( 'woocommerce_setup_jetpack_opted_in' );
 
 				$tos_accepted = true;
-			} else {
-				$tos_accepted = WC_Connect_Options::get_option( 'tos_accepted' );
 			}
 
 			add_action( 'admin_init', array( $this, 'admin_enqueue_scripts' ) );
